+++ src/main/java/com/swp/board/domain/Comment.java
@@ -11,7 +11,7 @@ public class Comment extends CreatedAtEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id", nullable = false)
-    private Long commentId;
+    private Integer commentId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name ="board_id", nullable = false)
